Backends : engines for p-value computation.

A backend is an S3 class which implments the methods pvals, cors and pvals_singleton. Different backends provide p-value computation under different scenarios (e.g. squared vs sum of squared correlations, Indpendence within Gene and SNP's) and may do complicated precomputations on the data matrices (X and Y) behind the scenes. Yet all of them provide a unified interface. The methods below are constructors for the differnt backend objects and return an S3 object which implements pvalue methods listed above.

backend.chisq(X, Y, parallel = FALSE, calc_full_cor = FALSE,
  fast_approx = FALSE)

backend.indepNormal(X, Y, calc_full_cor = FALSE)

backend.indepChiSq(X, Y, calc_full_cor = FALSE)

backend.normal(X, Y, calc_full_cor = FALSE)

backend.normal_two_sided(X, Y, calc_full_cor = FALSE)

backend.normalc(X, Y, parallel = FALSE, calc_full_cor = FALSE)

backend.perm(X, Y, calc_full_cor = FALSE)

Arguments

X
Matrix. The data vector for the X side

Y
Matrix. The data vector for the Y side

parallel
Logical Run pvalue computation using multiple threads?

calc_full_cor
Logical. Should it calculate the c(ncol(X),ncol(Y)) dimensional correlation matrix or not? Calculating this matrix upfront makes the pvalue computation faster but it also takes up lot of memory.

fast_approx
Use the BigX_fast approximation?

Functions

backend.chisq: Constructor for sum of squared correlations under uncorrelated Gene ans SNP sets (the weak null).

backend.indepNormal: Constructor for sum of correlations (one sided) under the strong independece (within Gene and Snips and between them).

backend.indepChiSq: Constructor for sum of squared correlations under strong independence.

backend.normal: Constructor for one sided sum of correlations under uncorrelated Gene ans SNP sets (the weak null).

backend.normal_two_sided: Constructor for two sided sum of correlations under uncorrelated Gene and SNP sets (the weak null).

backend.normalc: Constructor for sum of correlations under uncorrelated Gene ans SNP sets (the weak null).

backend.perm: Constructor for sum of squared correlations under independent Gene and SNP sets (the strong null).

Examples

# NOT RUN {
 X <- matrix(rnorm(100), nrow=10)
 Y <- matrix(rnorm(50), nrow=10)
 bk <- backend.indepNormal(X, Y)
 #Calculate Y pvalues againt the X (column) subset {1,2..,7}
 py <- pvals(bk, 1:7)
 #Calculate X pvalues againt the Y (column) subset {2, 3}. Use global numbering.
 px <- pvals(bk, 10 + c(2,3))
# }