Filter bimodules for results that look similar.

The bimodules are clustered based on their jaccard distances and the dendrogram is cut to obtain the same number of bimodules as are obtained by the effective-count formula. For each of the cut subtree an appropriate bimodule is chosen to represent it.

filter_bimodules(bms, hclust.method = "average", plot.dendrogram = FALSE)

Arguments

bms
a list of bimodules. Each bimodule should be a list with two elements x and y.

hclust.method
The clustering method to use (passed to hclust)

plot.dendrogram
logical Plot the dendrogram along with the line it is cut at.

Value

The list of filtered bimodules. If there are ties the first one in order is preferred.