Backends : engines for p-value computation.

A backend is an S3 class which implments the methods pvals, cors and pvals_singleton. Different backends provide p-value computation under different scenarios (e.g. squared vs sum of squared correlations, Indpendence within Gene and SNP's) and may do complicated precomputations on the data matrices (X and Y) behind the scenes. Yet all of them provide a unified interface. The methods below are constructors for the differnt backend objects and return an S3 object which implements pvalue methods listed above.

backend.perm(X, Y, cache.size = 0)

Arguments

X
Matrix. The data vector for the X side

Y
Matrix. The data vector for the Y side

cache.size
The cache size storing the correlations. Defaults to 0 (don't store anything).

Functions

backend.perm: Constructor for sum of squared correlations under independent Gene and SNP sets (the strong null).

Examples

# NOT RUN {
 X <- matrix(rnorm(100), nrow=10)
 Y <- matrix(rnorm(50), nrow=10)
 bk <- backend.indepNormal(X, Y)
 #Calculate Y pvalues againt the X (column) subset {1,2..,7}
 py <- pvals(bk, 1:7)
 #Calculate X pvalues againt the Y (column) subset {2, 3}. Use global numbering.
 px <- pvals(bk, 10 + c(2,3))
# }