Correlation Bi-community Extraction method

Given two groups of variables, find correlation bi-communities between them. For such a community, the nodes from the first group are are higly correlated to the community-nodes from the second group, and vice versa.

cbce(X, Y, alpha = 0.05, OL_thres = 0.9, break_thresh = 0.5,
  exhaustive = FALSE, OL_tol = Inf, Dud_tol = Inf,
  time_limit = Inf, updateMethod = 1,
  init_method = "conservative-BH", inv.length = TRUE,
  start_nodes = NULL, parallel = FALSE, calc_full_cor = FALSE,
  backend = "chisq", mask_extracted = FALSE,
  rank_initial_sets = FALSE, init_quick_update = FALSE,
  multiple_testing_method = "BHY", diagnostics = diagnostics1)

Arguments

X, Y
Numeric Matices. Represents the two groups of variables.

alpha
\(\in (0,1)\). Controls the type1 error per update. This is the type1 error to use for BH procedure

OL_thres
\(\in (0,1)\). Threshold used to conclude significant overlap. The sets have significant overlap if jaccard similarity is >= OL_thres.

break_thresh
Numeric A number between [0,1] that determines when a cycle will be broken. A value of 1 means that cycles will allways be broken and 0 that it will never be.

exhaustive
Boolean. If exhaustive is FALSE, new extractions are not started from nodes within found communities. Otherwise, attempt is made to start from all nodes.

OL_tol
If more than OL_tol found communities have OL_thres overlap, stop method.

Dud_tol
If more than Dud_tol initializations end in a dud, stop method.

time_limit
Stop method after time_limit seconds.

updateMethod
Use the 1(-step) update vs 2(-step) update

init_method
The initialization procedure to use. Must be one of "conservative-BH", "non-conservative-BH", "BH-0.5", "BH-0.5-nc", "BH-0.9-nc", "no-multiple-testing".

inv.length
Logical. Use inv.length as score while selecting the smallest community from communities with significant overlap.

start_nodes
The initial set of nodes to start with. If Null start from all the nodes (may still exclude nodes within found communities if exhaustive = FALSE).

parallel
Use parallel processing.

calc_full_cor
Calculate c(ncol(X),ncol(Y)) dimensional correlation matrix. This makes the computation faster but requires more memory.

backend
The engine to use for p-value computation. Currently must be one of "normal", "normalc", "normal_two_sided", "chisq", "indepChiSq", "indepNormal".

mask_extracted
Boolean. If true, then once a stable community is found, the nodes in it will never be used in future extractions.

rank_initial_sets
Logical Start extraction from inital sets with the higher scores.

init_quick_update
Logical Use a quick half-update in the init step.

multiple_testing_method
Method to use for multiple testing. Should be one of c('BHY', 'BH', 'sqrt_BH', 'sqrt_BHY', 'square_BH', 'Bonferroni')

diagnostics
This is a function that is called whenever internal events happen. It can then collect useful meta-data which is added to the final resutls.

Value

The return value is a list with details of the extraction and list of indices representing the communities. See example below (finding communities in noise). Note that the variables from the X and Y set are denoted using a single numbering. Hence the nodes in X are denoted by 1:dx and the nodes in Y are denoted by the numbers following dx (hence dx+1:dy)

Details

cbce applies an update function (mapping subsets of variables to subsets of variables) iteratively until a fixed point is found. These fixed points are reported as communities. The update function uses multiple-testing procedure to find variables correlated to the current set of variables.

The update starts from a single node (starting with the initialization step) and is repeated till either a fixed point is found or some set repeats. Each such run is called an extraction. Since the extraction only starts from singleton node, there are ncol(X)+ncol(Y) possible extractions.

Examples

# NOT RUN {
n <- 100
dx <- 50
dy <- 70

X <- matrix(rnorm(n*dx), ncol=dx)
Y <- matrix(rnorm(n*dy), ncol=dy)
res <- cbce(X, Y)
finalComms <- lapply(res$extract_res[res$finalIndxs], function(r) r$StableComm)
# }